refactor(command suggestion): map generated once
